update memory usage for dosage for qq plots

+++ workflow/scripts/common/utils.py
@@ -502,7 +502,7 @@ def get_ancestry_dosage(arr, n_anc):
 
 def load_true_la(path):
 	"""Load true local ancestry."""
-	return(np.load(path)['arr'].astype(np.half))
+	return(np.load(path)['arr'].astype(np.single))
 
 
 def get_true_anc_dosage(true_la, n_anc):
@@ -590,7 +590,8 @@ def load_rfmix_fb(path, sites_file):
 	# expand out to each site
 	# needed because RFMix2 only reports LA every fifth site.
 	rfmix_res = np.repeat(rfmix_res.iloc[:, 4:].values, [5], axis=0)
-	rfmix_res = rfmix_res.astype(np.half)
+	# rfmix_res = rfmix_res.astype(np.half)
+	rfmix_res = rfmix_res.astype(np.single)
 
 	nsites = len(pd.read_csv(sites_file, header=None))
 	rfmix_res = rfmix_res[:nsites, :]
@@ -603,7 +604,8 @@ def load_flare(path, sites_file, flare_sites, BCFTOOLS):
 	flare = flare.dropna(axis=1)
 	res = flare.iloc[:, 2:].values
 	res = np.concatenate([res[:1], res])
-	res = res.astype(np.half)
+	# res = res.astype(np.half)
+	res = res.astype(np.single)
 
 	# account for any sites filtered by flare (e.g. due to MAF)
 	pre_sites = pd.read_csv(sites_file, header=None).values.flatten()
@@ -619,7 +621,8 @@ def load_mosaic(path):
 	"""Return an array of the posterior LA probabilities from MOSAIC."""
 	arr = np.load(path)['arr']
 	res = arr.T.reshape((arr.shape[2], -1), order='C')
-	res = res.astype(np.half)
+	# res = res.astype(np.half)
+	res = res.astype(np.single)
 	return(res)
 
 
@@ -757,7 +760,7 @@ def make_qq_report(inferred_dosage, true_dosage, nbins):
 
 	qq['bin'] = pd.cut(
 		qq['inferred'],
-		bins=np.linspace(0, 2, nbins + 1),
+		bins=np.linspace(0, 2, nbins + 1, dtype="float32"),
 		include_lowest=True,
 		labels=False
 	).astype("int16")
